--- database/seeders/ItemTypeSeeder.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace Database\Seeders;
-
-use Illuminate\Database\Console\Seeds\WithoutModelEvents;
-use Illuminate\Database\Seeder;
-use App\Models\ItemType;
-
-class ItemTypeSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     */
-    public function run(): void
-    {
-        $types = ['default', 'shelve', 'light'];
-        foreach ($types as $type) {
-            ItemType::create(['name' => $type]);
-        }
-    }
-}
