--- database/seeders/WhtagSeeder.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace Database\Seeders;
-
-use App\Models\Whtag;
-use Illuminate\Database\Seeder;
-
-class WhtagSeeder extends Seeder
-{
-    public function run(): void
-    {
-        $tags = ['車', 'バイク', '棚', 'ライト'];
-        foreach ($tags as $tag) {
-            Whtag::create(['name' => $tag]);
-        }
-    }
-}
